# docs/swagger.yaml
basePath: /api/v1
definitions:
  handler.errorResponse:
    properties:
      error:
        type: string
    type: object
  model.AuthResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/model.User'
    type: object
  model.Category:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
  model.CategoryInput:
    properties:
      description:
        type: string
      name:
        maxLength: 100
        type: string
    required:
    - name
    type: object
...
  model.Invoice:
    properties:
      amount:
        type: number
      created_at:
        type: string
      id:
        type: integer
      issued_date:
        type: string
      number:
        type: string
      purchase_request_id:
        type: integer
      status:
        type: string
      updated_at:
        type: string
    type: object
  model.InvoiceInput:
    properties:
      amount:
        minimum: 0
        type: number
      number:
        maxLength: 50
        type: string
      purchase_request_id:
        type: integer
      status:
        enum:
        - unpaid
        - paid
        - void
        type: string
    required:
    - number
    type: object
  model.LoginInput:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
...
  model.RegisterInput:
    properties:
      password:
        maxLength: 72
        minLength: 6
        type: string
      role:
        enum:
        - admin
        - user
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - password
    - username
    type: object
...
  model.SalesAnalytics:
    properties:
      from:
        type: string
      invoice_count:
        type: integer
      paid_amount:
        type: number
      to:
        type: string
      total_amount:
        type: number
      unpaid_amount:
        type: number
    type: object
  model.StatusCount:
    properties:
      count:
        type: integer
      status:
        type: string
    type: object
  model.Summary:
    properties:
      total_categories:
        type: integer
      total_invoices:
        type: integer
      total_products:
        type: integer
      total_purchase_requests:
        type: integer
      total_revenue:
        type: number
      total_suppliers:
        type: integer
      total_users:
        type: integer
    type: object
...
  model.TopProduct:
    properties:
      name:
        type: string
      product_id:
        type: integer
      request_count:
        type: integer
      total_quantity:
        type: integer
    type: object
...
  model.User:
    properties:
      created_at:
        type: string
      id:
        type: integer
      role:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: CRM Backend
  description: |-
    Бэкенд CRM-системы для предприятия оптово-розничной торговли:
    управление продуктами, категориями, поставщиками, заявками на закупку,
    счетами-фактурами, аналитикой и отчётами.
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: CRM API
  version: "1.0"
paths:
  /analytics/purchase-requests:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.StatusCount'
            type: array
      security:
      - BearerAuth: []
      summary: Количество заявок по статусам
      tags:
      - analytics
  /analytics/sales:
    get:
      parameters:
      - description: Дата с (YYYY-MM-DD)
        in: query
        name: from
        type: string
      - description: Дата по (YYYY-MM-DD)
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SalesAnalytics'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Финансовая аналитика за период
      tags:
      - analytics
  /analytics/summary:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Summary'
      security:
      - BearerAuth: []
      summary: Сводные KPI
      tags:
      - analytics
  /analytics/top-products:
    get:
      parameters:
      - description: Сколько вернуть (по умолчанию 10)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.TopProduct'
            type: array
      security:
      - BearerAuth: []
      summary: Самые востребованные продукты
      tags:
      - analytics
  /auth/login:
    post:
      consumes:
      - application/json
      description: Проверяет учётные данные и возвращает JWT
      parameters:
      - description: Логин и пароль
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Вход в систему
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Создаёт нового пользователя с ролью user и возвращает JWT
      parameters:
      - description: Данные регистрации
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.RegisterInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Регистрация пользователя
      tags:
      - auth
  /categories:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Category'
            type: array
      security:
      - BearerAuth: []
      summary: Список категорий
      tags:
      - categories
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные категории
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.CategoryInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Создать категорию
      tags:
      - categories
  /categories/{id}:
    delete:
      parameters:
      - description: ID категории
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Удалить категорию
      tags:
      - categories
    get:
      parameters:
      - description: ID категории
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Category'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Получить категорию
      tags:
      - categories
    put:
      consumes:
      - application/json
      parameters:
      - description: ID категории
        in: path
        name: id
        required: true
        type: integer
      - description: Данные категории
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.CategoryInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Category'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Обновить категорию
      tags:
      - categories
  /invoices:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Invoice'
            type: array
      security:
      - BearerAuth: []
      summary: Список счетов-фактур
      tags:
      - invoices
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные счёта
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.InvoiceInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Invoice'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Создать счёт-фактуру
      tags:
      - invoices
  /invoices/{id}:
    delete:
      parameters:
      - description: ID счёта
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Удалить счёт-фактуру
      tags:
      - invoices
    get:
      parameters:
      - description: ID счёта
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Invoice'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Получить счёт-фактуру
      tags:
      - invoices
    put:
      consumes:
      - application/json
      parameters:
      - description: ID счёта
        in: path
        name: id
        required: true
        type: integer
      - description: Данные счёта
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.InvoiceInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Invoice'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Обновить счёт-фактуру
      tags:
      - invoices
  /me:
    get:
      description: Возвращает данные авторизованного пользователя
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Текущий пользователь
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: 'Введите токен в формате: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...